Redstone power should be an int, added GuiBlock, fixed typo in CommandBlock.

+++ src/main/java/org/spongepowered/api/block/Block.java
@@ -43,5 +43,5 @@ public interface Block {
      *
      * @return the block's redstone power level.
      */
-    float getPowerLevel();
+    int getPowerLevel();
 }
